refactor job
