+++ pkg/vm/engine/tae/db/open.go
@@ -250,21 +250,8 @@ func Open(
 
 	// Init timed scanner
 
-	db.MergeScheduler = merge.NewScheduler(db.Runtime, merge.NewTaskServiceGetter(opts.TaskServiceGetter))
-	scanner := newMergeScanner(db.Catalog, db.LogtailMgr, db.MergeScheduler)
-	db.MergeJob = tasks.NewCancelableCronJob(
-		"merge",
-		opts.CheckpointCfg.ScanInterval,
-		func(ctx context.Context) {
-			scanner.OnExec()
-		},
-		true,
-		1,
-	)
-
 	db.Wal.Start()
 	db.BGCheckpointRunner.Start()
-	db.MergeJob.Start()
 	// TODO: WithGCInterval requires configuration parameters
 	gc2.SetDeleteTimeout(opts.GCCfg.GCDeleteTimeout)
 	gc2.SetDeleteBatchSize(opts.GCCfg.GCDeleteBatchSize)
@@ -371,6 +358,18 @@ func Open(
 		},
 		1,
 	)
+
+	db.CronJobs.AddJob("Compact",
+		opts.CheckpointCfg.ScanInterval,
+		func(ctx context.Context) { db.LogtailMgr.TryCompactTable() },
+		0)
+
+	db.MergeScheduler = merge.NewScheduler(db.Runtime, merge.NewTaskServiceGetter(opts.TaskServiceGetter))
+	scanner := newMergeScanner(db.Catalog, db.MergeScheduler)
+	db.CronJobs.AddJob("Merge",
+		opts.CheckpointCfg.ScanInterval,
+		func(ctx context.Context) { scanner.OnExec() },
+		1)
 
 	if opts.CheckpointCfg.MetadataCheckInterval != 0 {
 		db.CronJobs.AddJob(
