+++ pkg/vm/engine/tae/tasks/handler.go
@@ -16,9 +16,11 @@ package tasks
 
 import (
 	"context"
+	"sync"
 
 	"github.com/matrixorigin/matrixone/pkg/common/moerr"
 	"github.com/matrixorigin/matrixone/pkg/logutil"
+	"github.com/panjf2000/ants/v2"
 )
 
 var (
@@ -48,4 +50,56 @@ func (h *baseTaskHandler) Enqueue(task Task) {
 
 func (h *baseTaskHandler) Execute(task Task) {
 	h.execFunc(task)
+}
+
+var (
+	poolHandlerName = "PoolHandler"
+)
+
+type poolHandler struct {
+	baseTaskHandler
+	opExec OpExecFunc
+	pool   *ants.Pool
+	wg     *sync.WaitGroup
+}
+
+func NewPoolHandler(ctx context.Context, num int) *poolHandler {
+	pool, err := ants.NewPool(num)
+	if err != nil {
+		panic(err)
+	}
+	h := &poolHandler{
+		baseTaskHandler: *NewBaseEventHandler(ctx, poolHandlerName),
+		pool:            pool,
+		wg:              &sync.WaitGroup{},
+	}
+	h.opExec = h.execFunc
+	h.execFunc = h.doHandle
+	return h
+}
+
+func (h *poolHandler) Execute(task Task) {
+	h.opExec(task)
+}
+
+func (h *poolHandler) doHandle(op IOp) {
+	closure := func(o IOp, wg *sync.WaitGroup) func() {
+		return func() {
+			h.opExec(o)
+			wg.Done()
+		}
+	}
+	h.wg.Add(1)
+	err := h.pool.Submit(closure(op, h.wg))
+	if err != nil {
+		logutil.Warnf("%v", err)
+		op.SetError(err)
+		h.wg.Done()
+	}
+}
+
+func (h *poolHandler) Stop() {
+	h.pool.Release()
+	h.baseTaskHandler.Stop()
+	h.wg.Wait()
 }
